+++ pyArango/query.py
@@ -144,9 +144,9 @@ class AQLQuery(Query) :
         request = self.connection.session.post(database.cursorsURL, data = json.dumps(payload, cls=json_encoder))
         Query.__init__(self, request, database, rawResults)
 
-    def explain(self, allPlans = False) :
+    def explain(self, bindVars={}, allPlans = False) :
         """Returns an explanation of the query. Setting allPlans to True will result in ArangoDB returning all possible plans. False returns only the optimal plan"""
-        return self.database.explainAQLQuery(self.query, allPlans)
+        return self.database.explainAQLQuery(self.query, bindVars, allPlans)
 
     def _raiseInitFailed(self, request) :
         data = request.json()